requirements.txt
asgiref==3.5.2
Django==4.1.2
django-tempus-dominus==5.1.2.17
django-widget-tweaks==1.4.12
sqlparse==0.4.3
